@@ -16,10 +16,10 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 # In Windows, build-name is used as the major, minor, and patch parts
 # of the product and file versions while build-number is used as the build suffix.
-version: 1.6.0+13
+version: 1.7.0+14
 
 environment:
-  sdk: '>=2.19.3 <3.0.0'
+  sdk: '>=3.3.0 <4.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -35,28 +35,28 @@ dependencies:
 
   cupertino_icons: ^1.0.2
   delayed_display: ^2.0.0
-  shared_preferences: ^2.0.18
-  sqflite: ^2.2.5
-  path_provider: ^2.0.13
-  path: ^1.8.2
-  chat_gpt_sdk: ^2.0.6
-  http: ^0.13.5
-  cached_network_image: ^3.2.3
+  shared_preferences: ^2.2.2
+  sqflite: ^2.3.2
+  path_provider: ^2.1.2
+  path: ^1.9.0
+  chat_gpt_sdk: ^3.0.4
+  http: ^1.2.1
+  cached_network_image: ^3.3.1
   loading_animation_widget: ^1.2.0+4
   flutter_launcher_icons: ^0.13.1
-  share_plus: ^7.0.2
-  url_launcher: ^6.1.11
-  fluttertoast: ^8.2.2
+  share_plus: ^7.2.2
+  url_launcher: ^6.2.5
+  fluttertoast: ^8.2.4
   sliding_up_panel: ^2.0.0+1
-  oktoast: ^3.3.1
-  tutorial_coach_mark: ^1.2.9
-  easy_localization: ^3.0.2
+  oktoast: ^3.4.0
+  tutorial_coach_mark: ^1.2.11
+  easy_localization: ^3.0.5
   toggle_switch: ^2.1.0
   auto_size_text: ^3.0.0
   firebase_analytics: ^10.6.3
   firebase_core: ^2.21.0
   in_app_review: ^2.0.8
-  google_mobile_ads: ^3.1.0
+  google_mobile_ads: ^4.0.0
 
 dev_dependencies:
   flutter_test:
