@@ -132,17 +132,17 @@ class _SettingsPageState extends State<SettingsPage> {
                     const SizedBox(height: 4),
                     Text(
                       'Watch next uses TMDb but is not endorsed or certified by TMDb.',
-                      style: Theme.of(context).textTheme.displaySmall,
+                      style: Theme.of(context).textTheme.labelSmall,
                     ),
                     const SizedBox(height: 4),
                     Text(
                       'Watch provider information is provided by JustWatch.',
-                      style: Theme.of(context).textTheme.displaySmall,
+                      style: Theme.of(context).textTheme.labelSmall,
                     ),
                     const SizedBox(height: 8),
                     Text(
                       'We do not claim ownership of any of the images or data provided.',
-                      style: Theme.of(context).textTheme.displaySmall,
+                      style: Theme.of(context).textTheme.labelSmall,
                     ),
                     const DividerWidget(padding: 0, height: 16),
                   ],
